Make user controller tests integration tests
Since the user "flow" is now distributed among different
controllers (the UsersController + Devise's RegistrationsController,
etc.), and since it already is technically an integration test, it
made sense to me to rename this the UserFlowsTest and put it under
test/integration. The alternative was to create a new test file for
the RegistrationsController just to test signup, which seemed less
appealing.