Remove users/alerts association
Users have many alerts, with a dependent destroy hook. Because
alerts don't have a user_id column, users can't be destroyed (it
results in a SQL error when looking for alerts with the corresponding
user id to destroy). Based on [this
comment](https://github.com/lbaillie/assemble-api/pull/2#issuecomment-275895345),
there shouldn't be a users-alerts relationship, so removing the
`has_many` is sufficient to allow us to destroy users again.